keep project membership and dependency rules in the SDK
