+++ packages/server/src/api/controllers/ai/agentOperations.ts
@@ -30,7 +30,8 @@ async function createAgentOperationUnlocked(
     allowKnowledgeSourceDownload: body.allowKnowledgeSourceDownload ?? true,
     escalation: body.escalation,
   })
-  await propagateProjectDependencyChangesWithWarning(ctx, {
+  await propagateProjectDependencyChangesWithWarning({
+    ctx,
     rootResourceId: agent._id!,
     currentProjectIds: agent.projectIds,
     previousProjectIds: existing.projectIds,
@@ -70,7 +71,8 @@ async function updateAgentOperationUnlocked(
 
   const existing = await sdk.ai.agents.getOrThrow(agentId)
   const agent = await sdk.ai.agents.updateOperation(agentId, operationId, body)
-  await propagateProjectDependencyChangesWithWarning(ctx, {
+  await propagateProjectDependencyChangesWithWarning({
+    ctx,
     rootResourceId: agent._id!,
     currentProjectIds: agent.projectIds,
     previousProjectIds: existing.projectIds,
