+++ packages/server/src/api/controllers/datasource.ts
@@ -41,11 +41,7 @@ import { getQueryParams, getTableParams } from "../../db/utils"
 import sdk from "../../sdk"
 import { processTable } from "../../sdk/workspace/tables/getters"
 import { invalidateCachedVariable } from "../../threads/utils"
-import {
-  propagateProjectDependencyChangesWithWarning,
-  resolveProjectIds,
-  resolveUpdatedProjectIds,
-} from "../../utilities/projects"
+import { propagateProjectDependencyChangesWithWarning } from "../../utilities/projects"
 import { builderSocket } from "../../websockets"
 
 async function clearOAuth2TokenCaches(datasource: Datasource) {
@@ -255,10 +251,10 @@ async function updateUnlocked(
     ...baseDatasource,
     ...sdk.datasources.mergeConfigs(dataSourceBody, baseDatasource),
   }
-  datasource.projectIds = await resolveUpdatedProjectIds(
-    ctx.request.body.projectIds,
-    baseDatasource.projectIds
-  )
+  datasource.projectIds = await sdk.projects.resolveUpdatedProjectIds({
+    projectIds: ctx.request.body.projectIds,
+    currentProjectIds: baseDatasource.projectIds,
+  })
   stripDatasourceEntityProjectIds(datasource)
 
   // this block is specific to GSheets, if no auth set, set it back
@@ -315,7 +311,8 @@ async function updateUnlocked(
     : await persistDatasource()
   datasource._rev = response.rev
 
-  await propagateProjectDependencyChangesWithWarning(ctx, {
+  await propagateProjectDependencyChangesWithWarning({
+    ctx,
     rootResourceId: datasource._id!,
     currentProjectIds: datasource.projectIds,
     previousProjectIds: baseDatasource.projectIds,
@@ -358,7 +355,9 @@ export async function save(
   )
 
   await sdk.projects.doWithProjectAssignmentsLockIfEnabled(async () => {
-    datasource.projectIds = await resolveProjectIds(datasource.projectIds)
+    datasource.projectIds = await sdk.projects.resolveProjectIds(
+      datasource.projectIds
+    )
     stripDatasourceEntityProjectIds(datasource)
     const persistDatasource = async () => {
       const restTemplateId = datasource.restTemplateId
@@ -382,7 +381,8 @@ export async function save(
       await persistDatasource()
     }
 
-    await propagateProjectDependencyChangesWithWarning(ctx, {
+    await propagateProjectDependencyChangesWithWarning({
+      ctx,
       rootResourceId: datasource._id!,
       currentProjectIds: datasource.projectIds,
       previousProjectIds: [],
